chat_service/requirements.txt
fastapi
uvicorn[standard]
pyjwt
uvicorn
python-jose
pymongo
python-dotenv
pika